chore: Update network configuration and hostnames
This commit updates the network configuration and hostnames in the /etc/hosts file. The IP addresses and hostnames have been changed to reflect the new naming convention. The base-vm is now b4s3, vm-master is now m4st3r, vm-1-worker is now w0rk3r-1, and vm-2-worker is now w0rk3r-2. These changes align with recent updates to the setup scripts and improve the clarity and consistency of the network configuration.